refactor: extract RelationshipEdge.css.ts
