+++ packages/diagram/src/overlays/edge-details/xyflow/RelationshipEdge.tsx
@@ -6,7 +6,7 @@ import { useDiagramState } from '../../../hooks/useDiagramState'
 import { stopPropagation } from '../../../xyflow/utils'
 import { useOverlayDialog } from '../../OverlayContext'
 import { ZIndexes } from '../use-layouted-edge-details'
-import * as css from './styles.css'
+import * as css from '../../shared/xyflow/RelationshipEdge.css'
 import { only } from 'remeda'
 import { type BaseTypes } from '../../shared/_types'
